Missing commit message that really applies to http://lists.mysql.com/commits/113698 :
This commit rewrites the mechanism for iterating over results for linked 
operations with multiple scans such that it uses a top-down rather than bottom 
up approach. We thus iterate linearly over the result tuples of the root 
operation and match these with corrsponding descendant tuples. The motivation 
for this change is:
* It simplifies handling of left outer join semantics.
* It (hopefully) simplifies the code in general.
* It should be easier to add support for "bushy" scan queries.

This commit contains a (preliminary) fix for for another problem: Non-root scan 
operations would produce duplicate tuple identifiers for clusters with more
than one data node. This has been fixed by including the data node identifier in
the tuple id.